(fix) 수익율 소수점 안되던 버그 수정

+++ src/main/java/step02/code/Main.java
@@ -34,7 +34,7 @@ public class Main {
       ResultView.result(i, grade.prize(), matched);
     }
 
-    float rate = total / money;
+    float rate = (float) total / money;
     ResultView.rate(rate);
     
     InputView.close();
